refactor: combine AnyAuthentication* and Authentication* classes
