+++ core/src/main/java/org/bitcoinj/wallet/AnyKeyChainGroup.java
@@ -39,12 +39,6 @@ import org.bitcoinj.script.ScriptBuilder;
 import org.bitcoinj.script.ScriptPattern;
 import org.bitcoinj.utils.ListenerRegistration;
 import org.bitcoinj.utils.Threading;
-import org.bitcoinj.wallet.bls.AnyBasicKeyChain;
-import org.bitcoinj.wallet.bls.AnyDefaultKeyChainFactory;
-import org.bitcoinj.wallet.bls.AnyDeterministicKeyChain;
-import org.bitcoinj.wallet.bls.IKeyBag;
-import org.bitcoinj.wallet.bls.AnyKeyChainFactory;
-import org.bitcoinj.wallet.bls.IRedeemData;
 import org.bitcoinj.wallet.listeners.KeyChainEventListener;
 import org.bouncycastle.crypto.params.KeyParameter;
 import org.slf4j.Logger;
