+++ core/src/main/java/org/bitcoinj/wallet/IRedeemData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.wallet.bls;
+package org.bitcoinj.wallet;
 
 import com.google.common.base.MoreObjects;
 import org.bitcoinj.crypto.IKey;
